Refactor out `lovatAPI.getCSVExport()`
