+++ lib/pages/settings.dart
@@ -10,6 +10,7 @@ import 'package:scouting_dashboard_app/reusable/friendly_error_view.dart';
 import 'package:scouting_dashboard_app/reusable/inset_picker.dart';
 import 'package:scouting_dashboard_app/reusable/lovat_api/delete_account.dart';
 import 'package:scouting_dashboard_app/reusable/lovat_api/get_analysts.dart';
+import 'package:scouting_dashboard_app/reusable/lovat_api/get_csv_export.dart';
 import 'package:scouting_dashboard_app/reusable/lovat_api/get_team_code.dart';
 import 'package:scouting_dashboard_app/reusable/lovat_api/get_tournaments.dart';
 import 'package:scouting_dashboard_app/reusable/lovat_api/get_user_profile.dart';
